pets/requirements.txt
asgiref==3.4.1
boto3==1.18.19
botocore==1.21.19
Django==3.2.6
django-storages==1.11.1
djangorestframework==3.12.4
djangorestframework-api-key==2.0.0
jmespath==0.10.0
Pillow==8.3.1
psycopg2==2.9.1
python-dateutil==2.8.2
python-dotenv==0.19.0
pytz==2021.1
s3transfer==0.5.0
six==1.16.0
sqlparse==0.4.1
urllib3==1.26.6
